+++ src/game_state/__init__.py
@@ -17,10 +17,16 @@ __copyright__ = "Copyright 2024-present Krish Mohan M."
 
 from typing import Literal, NamedTuple
 
-from .manager import StateManager
-from .state import State
+from .async_machine import AsyncState, AsyncStateManager
+from .sync_machine import State, StateManager
 
-__all__ = ("State", "StateManager", "version_info")
+__all__ = (
+    "AsyncState",
+    "AsyncStateManager",
+    "State",
+    "StateManager",
+    "version_info",
+)
 
 
 class VersionInfo(NamedTuple):