+++ src/game_state/sync_machine/__init__.py
@@ -1,0 +1,4 @@
+from .manager import StateManager
+from .state import State
+
+__all__ = ("State", "StateManager")