+++ tests/test_async/test_hooks.py
@@ -1,0 +1,42 @@
+from __future__ import annotations
+
+from typing import Any
+
+import pytest
+
+from src.game_state import AsyncState, AsyncStateManager
+from src.game_state.errors import StateError
+
+
+@pytest.mark.asyncio
+async def test_hooks() -> None:
+    state_manager = AsyncStateManager[AsyncState[Any]]()
+
+    await state_manager.connect_state_hook(
+        "tests.test_async.test_hooks_states.hook_1"
+    )
+    await state_manager.connect_state_hook(
+        "tests.test_async.test_hooks_states.hook_2"
+    )
+
+    STATE_1_NAME = "HookState1"
+    TOTAL_VALID_LOADED_STATES: int = 2
+
+    await state_manager.change_state(STATE_1_NAME)
+
+    assert state_manager.current_state is not None, "Expected non-None value."
+
+    assert state_manager.current_state.state_name == STATE_1_NAME, (
+        f"Expected `{STATE_1_NAME}` as current state, instead got {state_manager.current_state}"
+    )
+
+    with pytest.raises(StateError):
+        await state_manager.connect_state_hook(
+            "tests.test_async.test_hooks_states.hook_3"
+        )
+
+    loaded_states = len(state_manager.state_map)
+
+    assert loaded_states == TOTAL_VALID_LOADED_STATES, (
+        f"Expected {TOTAL_VALID_LOADED_STATES} loaded states, insted got {loaded_states}"
+    )